refactor: Station에 대한 Controller, Service 삭제
