--- src/main/java/subway/domain/station/dto/StationCreateRequest.java
@@ -1,16 +1,0 @@
-package subway.domain.station.dto;
-
-public class StationCreateRequest {
-    private String stationName;
-
-    public StationCreateRequest() {
-    }
-
-    public StationCreateRequest(String stationName) {
-        this.stationName = stationName;
-    }
-
-    public String getStationName() {
-        return stationName;
-    }
-}
