--- src/main/java/subway/ui/StationController.java
@@ -1,32 +1,0 @@
-package subway.ui;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.domain.station.Station;
-import subway.domain.station.dto.StationCreateRequest;
-import subway.domain.station.dto.StationCreateResponse;
-import subway.service.StationService;
-
-import java.net.URI;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-    private final StationService stationService;
-
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<StationCreateResponse> createStation(@RequestBody StationCreateRequest createRequest) {
-        Station createStation = stationService.saveStation(createRequest);
-        StationCreateResponse stationCreateResponse = new StationCreateResponse(createStation.getId(),
-                createStation.getName());
-        return ResponseEntity.created(URI.create("/stations/" + stationCreateResponse.getStationId()))
-                .body(stationCreateResponse);
-    }
-}
