+++ src/main/java/subway/ui/line/dto/AddStationToLineResponse.java
@@ -1,13 +1,13 @@
-package subway.domain.line.dto;
+package subway.ui.line.dto;
 
 import java.util.List;
 
-public class DeleteStationFromLineResponse {
+public class AddStationToLineResponse {
     private final Long lineId;
     private final String lineName;
     private final List<Long> stationIds;
 
-    public DeleteStationFromLineResponse(Long lineId, String lineName, List<Long> stationIds) {
+    public AddStationToLineResponse(Long lineId, String lineName, List<Long> stationIds) {
         this.lineId = lineId;
         this.lineName = lineName;
         this.stationIds = stationIds;
