+++ src/main/java/subway/ui/line/dto/LineCreateRequest.java
@@ -1,4 +1,4 @@
-package subway.domain.line.dto;
+package subway.ui.line.dto;
 
 public class LineCreateRequest {
     private final String lineName;
