+++ src/test/java/subway/integration/AddStationToLineTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.domain.line.Direction;
-import subway.domain.line.dto.AddStationToLineRequest;
-import subway.domain.line.dto.LineCreateRequest;
+import subway.ui.line.dto.AddStationToLineRequest;
+import subway.ui.line.dto.LineCreateRequest;
 
 import static org.hamcrest.Matchers.equalTo;
